--- precompiles/ibc/legacy/v520/abi.json
@@ -1,56 +1,0 @@
-[
-    {
-      "inputs": [
-        {
-          "internalType": "string",
-          "name": "toAddress",
-          "type": "string"
-        },
-        {
-          "internalType": "string",
-          "name": "port",
-          "type": "string"
-        },
-        {
-          "internalType": "string",
-          "name": "channel",
-          "type": "string"
-        },
-        {
-          "internalType": "string",
-          "name": "denom",
-          "type": "string"
-        },
-        {
-          "internalType": "uint256",
-          "name": "amount",
-          "type": "uint256"
-        },
-        {
-          "internalType": "uint64",
-          "name": "revisionNumber",
-          "type": "uint64"
-        },
-        {
-          "internalType": "uint64",
-          "name": "revisionHeight",
-          "type": "uint64"
-        },
-        {
-          "internalType": "uint64",
-          "name": "timeoutTimestamp",
-          "type": "uint64"
-        }
-      ],
-      "name": "transfer",
-      "outputs": [
-        {
-          "internalType": "bool",
-          "name": "success",
-          "type": "bool"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    }
-  ]
