+++ precompiles/pointer/legacy/v552/abi.json
